+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 Not released
 ------------
 
+* Change CI build to require x86-64-v3 instead of just AVX2. This is roughly
+  equivalent in terms of supported CPUs, but allows more instructions to be
+  used.
+
 v0.6.0 (2024/09/03)
 -------------------